Fix header extension inconsistency
