--- Extensions/Template/Extension.h
@@ -1,145 +1,0 @@
-/* Extension.h
- * This is where you define your extension
- * class. Runtime variables and function
- * declarations go in here; A/C/E function
- * definitions go in their respective .cpp
- * files, other function definitions go in
- * Extension.cpp. This header is included
- * at the bottom of Common.h, so you have
- * access to everything included by it.
- */
-
-class Extension
-{
-public:
-	/* rd
-	 * This is a pointer to the classic
-	 * RunData structure. Some things still
-	 * have to be stored there for MMF2's
-	 * sake, so you need to have acces to
-	 * that data.
-	 */
-	RD *rd;
-	/* rh
-	 * This is a pointer to the RunHeader
-	 * structure, which contains some
-	 * application information.
-	 */
-	RunHeader *rh;
-
-	/* Runtime
-	 * This is EDIF's compendium of useful
-	 * functions for every day life as an
-	 * extension. You can trigger immediate
-	 * events, read and write global data,
-	 * etc. See the EDIF Help file for
-	 * more information.
-	 */
-	Edif::Runtime Runtime;
-
-	/* MinimumBuild
-	 * Always set this to the latest build
-	 * of MMF2 that has been released, unless
-	 * you need to set it higher for a beta
-	 * version of MMF2 to get access to some
-	 * new SDK feature.
-	 */
-	static const int MinimumBuild = 255;
-	/* Version
-	 * This is the version of your extension
-	 * that is stored in the editdata. You should
-	 * only change this when your editdata structure
-	 * changes, not just when you have a new version
-	 * of your extension with new features.
-	 */
-	static const int Version = 1;
-
-	/* OEFLAGS, OEPREFS
-	 * These are settings for your extension
-	 * that tell MMF2 how to treat your extension
-	 * and what behaviors and responsibilities
-	 * to add. Check the MMF2SDK Help file for
-	 * more detailed information on the flags you
-	 * can set and what they implicate.
-	 */
-	static const int OEFLAGS = OEFLAG_VALUES|OEFLAG_RUNBEFOREFADEIN|OEFLAG_NEVERKILL;
-	static const int OEPREFS = 0;
-
-	/* WindowProcPrioerity
-	 * If your extension creates a Window
-	 * and implements the WindowProc function,
-	 * this is the priority for that stuff.
-	 */
-	static const int WindowProcPriority = 100;
-
-	Extension(RD *rd, SerializedED *SED, createObjectInfo *COB);	//defined & documented in Extension.cpp
-	~Extension();													//defined & documented in Extension.cpp
-
-
-	/* Add any data you want to store in your
-	 * extension to this class (eg. what you'd
-	 * normally store in the rundata). Unlike
-	 * other SDKs, you can store real C++
-	 * objects with constructors and destructors,
-	 * without having to call them manually or
-	 * store a pointer.
-	 */
-
-	//stdtstring MyString;
-	//int MyInt;
-	//std::vector<float> MyArray;
-
-
-	/* Add your actions, conditions and expressions
-	 * as real class member functions here. The
-	 * arguments (and return types for expressions/conditions)
-	 * must match EXACTLY what you defined in the JSON.
-	 * As a handy tip, all parameters will be either
-	 * a pointer, an int, or a float (each only 4 bytes)
-	 * and the same applies to return types. Remember to
-	 * link the actions, conditions and expressions to their
-	 * numeric IDs in the class constructor in Extension.cpp.
-	 */
-
-	//Actions
-	void ActionExample(int ExampleParameter);
-	void SecondActionExample();
-
-	//Conditions
-	bool AreTwoNumbersEqual(int FirstNumber, int SecondNumber);
-
-	//Expressions
-	int Add(int FirstNumber, int SecondNumber);
-	const char * HelloWorld();
-
-
-	/* Action, Condition, Expression
-	 * These are called if there's no function linked
-	 * to an ID. You may want to put MessageBox calls
-	 * to let you know that the ID is unlinked, or you
-	 * may just want to use unlinked A/C/Es as a feature.
-	 */
-
-	void Action(int ID, RD *rd, long param1, long param2);
-	long Condition(int ID, RD *rd, long param1, long param2);
-	long Expression(int ID, RD *rd, long param);
-
-
-
-
-	/* Handle, Display, Pause, Continue, Save, Load
-	 * These replace the functions like HandleRunObject
-	 * that used to be implemented in Runtime.cpp. They
-	 * work exactly the same, but they're inside the
-	 * extension class.
-	 */
-	short Handle();
-	short Display();
-
-	short Pause();
-	short Continue();
-
-	bool Save(HANDLE File);
-	bool Load(HANDLE File);
-
-};
